Rename to `index-animated`
