+++ src/index-animated.js
